type return type from attachTrace and separate getGcpProject id for peformance
