+++ lib/traceparent.ts
@@ -1,12 +1,10 @@
 import crypto from "crypto";
 
-export type Trace =
-  | {
-      traceId: string;
-      parentId: string;
-      isSampled: boolean;
-    }
-  | undefined;
+export type Trace = {
+  traceId: string;
+  parentId: string;
+  isSampled: boolean;
+};
 
 /**
  * Traceparent: 00-4bf92f3577b34da6a3ce929d0e0e4736-00f067aa0ba902b7-00
@@ -21,7 +19,7 @@ function traceIsSampled(traceFlags: string) {
   return (parseInt(traceFlags, 16) & FLAG_SAMPLED) === FLAG_SAMPLED;
 }
 
-export function getTraceFromTraceparent(traceHeader: string): Trace {
+export function getTraceFromTraceparent(traceHeader: string): Trace | undefined {
   const parts = traceHeader.split("-");
 
   if (!parts || parts.length !== 4) return;
